move cloud source repos out to its own package.
feel free to disagree on the final source/repos package name.
