+++ google-cloud-tools-plugin/src/com/google/cloud/tools/intellij/resources/ResourceErrorModelItem.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2016 The Android Open Source Project
+ * Copyright 2018 Google Inc. All Rights Reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -17,6 +17,21 @@
 package com.google.cloud.tools.intellij.resources;
 
 import javax.swing.tree.DefaultMutableTreeNode;
+import org.jetbrains.annotations.NotNull;
 
-/** This model item is shown when a resource manager call is outstanding. */
-public class ResourceLoadingModelItem extends DefaultMutableTreeNode {}
+/**
+ * This model item represents a node shown when an error occurs attempting to query a resource
+ * manager.
+ */
+public class ResourceErrorModelItem extends DefaultMutableTreeNode {
+
+  private String errorMessage;
+
+  public ResourceErrorModelItem(@NotNull String errorMessage) {
+    this.errorMessage = errorMessage;
+  }
+
+  public String getErrorMessage() {
+    return errorMessage;
+  }
+}
