+++ google-cloud-tools-plugin/src/com/google/cloud/tools/intellij/source/repos/ProjectRepositoriesModelItem.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2017 Google Inc. All Rights Reserved.
+ * Copyright 2018 Google Inc. All Rights Reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,12 +14,13 @@
  * limitations under the License.
  */
 
-package com.google.cloud.tools.intellij.resources;
+package com.google.cloud.tools.intellij.source.repos;
 
 import com.google.api.services.source.model.Repo;
 import com.google.cloud.tools.intellij.login.CredentialedUser;
+import com.google.cloud.tools.intellij.resources.ResourceEmptyModelItem;
+import com.google.cloud.tools.intellij.resources.ResourceErrorModelItem;
 import com.google.cloud.tools.intellij.util.GctBundle;
-import com.google.cloud.tools.intellij.vcs.CloudRepositoryService;
 import com.intellij.openapi.components.ServiceManager;
 import java.util.List;
 import javax.swing.tree.DefaultMutableTreeNode;
