+++ google-cloud-tools-plugin/testSrc/com/google/cloud/tools/intellij/source/repos/GcpHttpAuthProviderTest.java.excluded
